ai-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
groq==0.4.1
qdrant-client==1.7.0
sentence-transformers==2.2.2
psycopg2-binary==2.9.9
redis==5.0.1
python-dotenv==1.0.0
pydantic==2.5.2